requirements.txt
# Matrix client
matrix-nio[e2e]>=0.21.2

# Required for Matrix E2EE
python-olm>=3.2.15

# LLM backends (install at least one)
# Option A: llama-cpp-python (recommended for most users — works on CPU and GPU)
# llama-cpp-python>=0.2.56

# Option B: HuggingFace Transformers (uncomment if you prefer HF models)
transformers>=4.40.0
torch>=2.2.0
accelerate>=0.29.0
bitsandbytes>=0.43.0  # 4-bit / 8-bit quantization support

# Configuration & utilities
pyyaml>=6.0.1
schedule>=1.2.1
aiofiles>=23.2.1
aiohttp>=3.9.3
python-dateutil>=2.9.0
colorlog>=6.8.2

# Setup wizard
questionary>=2.0.1
rich>=13.7.1
